Rename `transport` module to `lsps0`
